# Improve faker.random.arrayElements performance