--- a/lib/random.js
+++ b/lib/random.js
@@ -82,14 +82,20 @@ function Random (faker, seed) {
         count = 0;
       }
 
-      var arrayCopy = array.slice();
-      var countToRemove = arrayCopy.length - count;
-      for (var i = 0; i < countToRemove; i++) {
-        var indexToRemove = faker.datatype.number({ max: arrayCopy.length - 1 });
-        arrayCopy.splice(indexToRemove, 1);
+      var arrayCopy = array.slice(0);
+      var i = array.length;
+      var min = i - count;
+      var temp;
+      var index;
+
+      while (i-- > min) {
+        index = Math.floor((i + 1) * faker.datatype.float({ min: 0, max: 1 }));
+        temp = arrayCopy[index];
+        arrayCopy[index] = arrayCopy[i];
+        arrayCopy[i] = temp;
       }
 
-      return arrayCopy;
+      return arrayCopy.slice(min);
   };
 
   /**