--- a/lib/random.js
+++ b/lib/random.js
@@ -89,7 +89,7 @@ function Random (faker, seed) {
       var index;
 
       while (i-- > min) {
-        index = Math.floor((i + 1) * faker.datatype.float({ min: 0, max: 1 }));
+        index = Math.floor((i + 1) * faker.datatype.float({ min: 0, max: 0.99 }));
         temp = arrayCopy[index];
         arrayCopy[index] = arrayCopy[i];
         arrayCopy[i] = temp;