`faker.random.arrayElements` has performance issues when dealing with large arrays.

The following implementation significantly improves performance.

Performance tests for an array of 200000 items:
| Count  | Old implementation | New implementation |
| -      | -                  | -                  |
| 20000  |      1872ms        |       10ms         |
| 40000  |      1643ms        |       9ms          |
| 60000  |      1542ms        |       9ms          |
| 80000  |      1429ms        |       10ms         |
| 100000 |      1270ms        |       12ms         |
| 120000 |      1099ms        |       15ms         |
| 140000 |      1013ms        |       25ms         |
| 160000 |      623ms         |       23ms         |
| 180000 |      333ms         |       22ms         |

Performance tests source code is available [here](https://codesandbox.io/s/faker-performance-test-oxsm4?file=/src/index.js)