@Marak 2021-03-18
Very nice @OleksandrYehorov 

Thank you for your contribution!

Could you share with us how you are running these benchmarks? We've been looking to implement benchmark testing for the library.