@OleksandrYehorov 2021-03-18
@Marak Here is the source code of the tests https://codesandbox.io/s/oxsm4

However, I don't think that my tests would be applicable for the whole library.