@makepanic 2021-03-19
This seems to change the behavior of arrayElements.
I wanted to see how it behaves when plotted using random rgb pixels.
I generate an array with the numbers 0..255 and use random.arrayElements with the length 3 to get each red green and blue value.

https://gist.github.com/makepanic/8b231c933707fe853213cd18275d4920

master:
![master](https://user-images.githubusercontent.com/1205444/111764932-0f70d900-88a4-11eb-99ff-69bf0608d827.png)

1143:
![1143](https://user-images.githubusercontent.com/1205444/111764949-14ce2380-88a4-11eb-8685-9633f4ac5914.png)

It seems like the random selection on master might be broken and this PR fixes it? Previously there were only blue hues. Now it's somewhat random.

Note: I might have made mistakes in the script so please verify it yourself :)