@OleksandrYehorov 2021-03-19
@makepanic Interesting observation! I just figured that while the implementation from master preserves the order of elements in the returned array, the new implementation does not.

```javascript
const array = [1, 2, 3, 4, 5, 6];
```
master:
```javascript
faker.random.arrayElements([1, 2, 3], 3); // Always returns [1, 2, 3]
```

PR:
```javascript
faker.random.arrayElements([1, 2, 3], 3); // Can return [2, 3, 1] [1, 3, 2] [1, 2, 3] etc.
```