@makepanic 2021-03-19
I'm seeing some cases in the generation of this PR returns an undefined value, e.g.:

![20210319-113730](https://user-images.githubusercontent.com/1205444/111767868-80fe5680-88a7-11eb-927e-d24bb90403ba.png)