@Marak 2021-03-20
Thank you for this additional review @makepanic 

The images you generated are super cool! I wonder if we can apply this to the library as a more generic method that would work for other `faker` API methods.

Merged.